chore: refactor databars components (#690)
