+++ src/apps/WorkOrder/components/index.ts
@@ -1,3 +1,1 @@
 export * from './FollowUpStatusFilter';
-export * from './EstimateBar';
-export * from './ExpendedProgressBar';
