+++ src/packages/Table/Components/Cells/Databars/ProgressBarContainer.styles.ts
